python/requirements.txt
uagents==0.6.2
ai_engine==0.1.3
pydantic==2.9.0
python-dotenv==1.0.0
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
openai>=1.0.0
